Remove fixtures and bundle component unit tests
