--- public/example-responses/postcode-UB7XNV.json
@@ -1,185 +1,0 @@
-{
-  "address_picker": false,
-  "addresses": [],
-  "dates": [
-    {
-      "date": "2019-12-12",
-      "polling_station": {
-        "polling_station_known": false,
-        "custom_finder": null,
-        "report_problem_url": null,
-        "station": null
-      },
-      "notifications": [],
-      "ballots": [
-        {
-          "ballot_paper_id": "parl.uxbridge-and-south-ruislip.2019-12-12",
-          "ballot_title": "Uxbridge and South Ruislip",
-          "poll_open_date": "2019-12-12",
-          "elected_role": "Member of Parliament",
-          "metadata": null,
-          "cancelled": false,
-          "replaced_by": null,
-          "replaces": null,
-          "ballot_url": "https://developers.democracyclub.org.uk/api/v1/elections/parl.uxbridge-and-south-ruislip.2019-12-12/",
-          "election_id": "parl.2019-12-12",
-          "election_name": "UK Parliament elections",
-          "post_name": "Uxbridge and South Ruislip",
-          "candidates_verified": false,
-          "candidates": [
-            {
-              "list_position": null,
-              "party": { "party_id": "ynmp-party:2", "party_name": "Independent" },
-              "person": {
-                "ynr_id": 72368,
-                "name": "Alfie John Utting",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72368/alfie-john-utting",
-                "email": null,
-                "photo_url": null
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "party:53", "party_name": "Labour Party" },
-              "person": {
-                "ynr_id": 47463,
-                "name": "Ali Reza Milani",
-                "absolute_url": "https://whocanivotefor.co.uk/person/47463/ali-reza-milani",
-                "email": "Ali@AliMilani.com",
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/dd/32/dd32b627edb55204641bfbd558161254.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "joint-party:53-119", "party_name": "Labour Party (joint Co-op)" },
-              "person": {
-                "ynr_id": 7292,
-                "name": "Bobby \"Elmo\" Smith",
-                "absolute_url": "https://whocanivotefor.co.uk/person/7292/bobby-elmo-smith",
-                "email": "givemebackelmo@outlook.com",
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/bf/a8/bfa88a4a561f6d3fb026e3248255d430.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "party:52", "party_name": "Conservative and Unionist Party" },
-              "person": {
-                "ynr_id": 4546,
-                "name": "Boris Johnson",
-                "absolute_url": "https://whocanivotefor.co.uk/person/4546/boris-johnson",
-                "email": "admin@hillingdonconservatives.org",
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/16/f9/16f919fe5dee6f997cc9575dc8a23ab4.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "ynmp-party:2", "party_name": "Independent" },
-              "person": {
-                "ynr_id": 72363,
-                "name": "Count Binface",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72363/count-binface",
-                "email": null,
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/18/eb/18eb26d2f07e4275ba247371e35d8eb5.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "party:85", "party_name": "UK Independence Party (UKIP)" },
-              "person": {
-                "ynr_id": 72366,
-                "name": "Geoffrey Denis Courtenay",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72366/geoffrey-denis-courtenay",
-                "email": null,
-                "photo_url": null
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "party:90", "party_name": "Liberal Democrats" },
-              "person": {
-                "ynr_id": 41972,
-                "name": "Jo Humphreys",
-                "absolute_url": "https://whocanivotefor.co.uk/person/41972/jo-humphreys",
-                "email": null,
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/5e/87/5e87094933f0c19986c50ec914fc5405.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": {
-                "party_id": "party:66",
-                "party_name": "Official Monster Raving Loony Party"
-              },
-              "person": {
-                "ynr_id": 72364,
-                "name": "Lord Buckethead",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72364/lord-buckethead",
-                "email": null,
-                "photo_url": null
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "party:63", "party_name": "Green Party" },
-              "person": {
-                "ynr_id": 34481,
-                "name": "Mark Keir",
-                "absolute_url": "https://whocanivotefor.co.uk/person/34481/mark-keir",
-                "email": "Mark.Keir@hillingdon.greenparty.org.uk",
-                "photo_url": "https://static-candidates.democracyclub.org.uk/media/cache/b7/73/b773c679a771a6363c04a76f40c29a1d.jpg"
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "ynmp-party:2", "party_name": "Independent" },
-              "person": {
-                "ynr_id": 72365,
-                "name": "Norma Burke",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72365/norma-burke",
-                "email": null,
-                "photo_url": null
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "ynmp-party:2", "party_name": "Independent" },
-              "person": {
-                "ynr_id": 72367,
-                "name": "William John Tobin",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72367/william-john-tobin",
-                "email": null,
-                "photo_url": null
-              }
-            },
-            {
-              "list_position": null,
-              "party": { "party_id": "ynmp-party:2", "party_name": "Independent" },
-              "person": {
-                "ynr_id": 72369,
-                "name": "Yace \"Interplanetary Time Lord\" Yogenstein",
-                "absolute_url": "https://whocanivotefor.co.uk/person/72369/yace-interplanetary-time-lord-yogenstein",
-                "email": null,
-                "photo_url": null
-              }
-            }
-          ],
-          "wcivf_url": "https://whocanivotefor.co.uk/elections/parl.uxbridge-and-south-ruislip.2019-12-12/uxbridge-and-south-ruislip/"
-        }
-      ]
-    }
-  ],
-  "electoral_services": {
-    "council_id": "E09000017",
-    "name": "London Borough of Hillingdon",
-    "email": "electoralservices@hillingdon.gov.uk",
-    "phone": "01895 250251",
-    "website": "http://www.hillingdon.gov.uk/",
-    "postcode": "UB8 1UW",
-    "address": "Electoral Registration Officer\nPhase II\nCivic Centre\nHigh Street; Uxbridge"
-  },
-  "postcode_location": {
-    "type": "Feature",
-    "properties": null,
-    "geometry": { "type": "Point", "coordinates": [-0.4709574285714285, 51.51086539999999] }
-  }
-}
