+++ public/index.html
@@ -69,7 +69,7 @@
         </pre>
         <p>Alternatively, the widget will pick up a postcode query parameter from the URL:</p>
         <pre>
-  https://domain.com/?postcode=UB78FA        
+  https://domain.com/?postcode=AA12AA        
         </pre>
         <p>For testing purposes, you can set a postcode once so that the dev server's auto-refresh always shows you the ballots/candidates section of the widget:</p>
         <form method="get" action="/">
