+++ src/tests/utils/common-responses.js
@@ -1,0 +1,79 @@
+export const listBallot = {
+  ballot_paper_id: 'gla.a.2020-05-07',
+  election_id: 'gla.a.2020-05-07',
+  candidates: [
+    {
+      list_position: 9,
+      party: { party_id: 'party:90', party_name: 'Liberal Democrats' },
+      person: {},
+    },
+    {
+      list_position: null,
+      party: { party_id: 'party:2755', party_name: "Women's Equality Party" },
+      person: {},
+    },
+    {
+      list_position: 5,
+      party: { party_id: 'party:90', party_name: 'Liberal Democrats' },
+      person: {},
+    },
+    {
+      list_position: 1,
+      party: { party_id: 'party:90', party_name: 'Liberal Democrats' },
+      person: {},
+    },
+    {
+      list_position: null,
+      party: { party_id: 'party:2755', party_name: "Women's Equality Party" },
+      person: {},
+    },
+  ],
+};
+
+export const normalBallot = {
+  ballot_paper_id: 'local.westminster.lancaster-gate.by.2018-11-22',
+  ballot_title: 'Westminster local election Lancaster Gate by-election',
+  election_id: 'local.westminster.2018-11-22',
+  election_name: 'Westminster local election',
+  post_name: 'Lancaster Gate',
+  candidates_verified: true,
+  candidates: [
+    {
+      list_position: null,
+      party: {
+        party_id: 'party:90',
+        party_name: 'Liberal Democrats',
+      },
+      person: {
+        ynr_id: 42252,
+        name: 'Sally Elizabeth Gray',
+        absolute_url: 'https://whocanivotefor.co.uk/person/42252/sally-elizabeth-gray',
+      },
+    },
+    {
+      list_position: null,
+      party: {
+        party_id: 'party:63',
+        party_name: 'Green Party',
+      },
+      person: {
+        ynr_id: 7505,
+        name: 'Zack Polanski',
+        absolute_url: 'https://whocanivotefor.co.uk/person/7505/zack-polanski',
+      },
+    },
+  ],
+};
+
+export const candidate = {
+  list_position: null,
+  party: { party_id: 'ynmp-party:2', party_name: 'Independent' },
+  person: {
+    ynr_id: 72363,
+    name: 'Example Name',
+    absolute_url: 'https://whocanivotefor.co.uk/person/72363/count-binface',
+    email: null,
+    photo_url:
+      'https://static-candidates.democracyclub.org.uk/media/cache/18/eb/18eb26d2f07e4275ba247371e35d8eb5.jpg',
+  },
+};
